pyproject.toml
[tool.poetry]
name = "mce-spec-inference"
version = "0.1.0"
description = "Maximum Causal Entropy Specification Inference."
authors = ["Marcell Vazquez-Chanlatte <mvc@linux.com>"]
license = "MIT"
readme="README.md"
repository = "https://github.com/mvcisback/mce-spec-inference"
packages = [
    { include = "mce"},
]

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^19.3"
bdd2dfa = "^0.4.0"
bidict = "^0.18.3"
dfa = {version = "^2.0.0", extras = ["draw"]}
funcy = "^1.13"
multiprocess = "^0.70.9"
py-aiger-bdd = "^0.3.1"
py-aiger-bv = "^2.0.0"
py-aiger-coins = {version = "^1.4.0", extras = ["sat", "bdd"]}
py-aiger-sat = "^1.1.0"
scipy = "^1.4.1"

[tool.poetry.dev-dependencies]
blessings = "^1.7"
codecov = "^2.0"
hypothesis = "^4.42"
ipython = "^7.13.0"
jupyter = {extras = ["notebook"], version = "^1.0.0"}
matplotlib = "^3.2.1"
notebook = "^6.0.3"
py-aiger-gridworld = "^0.2.0"
py-aiger-ptltl = "^1.2.0"
pygraphviz = "^1.5"
pytest-cov = "^2.7"
pytest-flake8 = "^1.0"
pytest-xdist = "^1.29"
pyyaml = "^5.3.1"
seaborn = "^0.10.0"
termplotlib = "^0.2.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
